requirements.txt
ultralytics
opencv-python
Pillow
pyserial
